requirements.txt
streamlit
pillow
pandas
numpy
mysqlclient
SQLAlchemy
